- fixed some uninitialized stuff in particle rendering. - added gl_ClampLight to all places where the light level is set.

+++ src/gl/scene/gl_wall.h
@@ -314,7 +314,7 @@ public:
 	friend struct GLDrawList;
 	friend void Mod_RenderModel(GLSprite * spr, model_t * mdl, int framenumber);
 
-	BYTE lightlevel;
+	int lightlevel;
 	BYTE foglevel;
 	BYTE hw_styleflags;
 	bool fullbright;
